replace ratatui with egui for emulation
The terminal only receives key down events, making it impossible to provide a
responsive experience when something should happen upon key release.

While the tui emulator is not recommended anymore, it's still possible to use
it with `--features tui`.

Egui fixes this, but it comes with the downside that the rendering loop must be
run on the main thread. For the library API to stay the way it is, and allow
users to control the rendering lifecycle themselves, the user application code
must be moved to a separate thread.

In my estimation, this doesn't introduce much complexity on the user's side. It
may just be a little confusing why it was done this way without knowing the
context. Users now have to use the library like this:

```rust
fn main() -> ! { led_matrix::run(app) }
fn app(matrix: &mut dyn led_matrix::LedMatrix) { /* user logic */ }
```